backend/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distlib==0.3.9
executing==2.2.0
factory_boy==3.3.3
Faker==37.4.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google==3.0.0
google-auth==2.40.3
google-genai==1.26.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy
packaging==25.0
pandas==2.3.0
pandas_ta==0.3.14b0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-cov>=6.2.0
pytest-subtests==0.14.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
ta==0.11.0
tenacity==8.5.0
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
vaderSentiment==3.3.2
virtualenv==20.31.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
marshmallow-dataclass==8.7.1
emoji==2.14.1
langdetect
google-api-python-client
ta>=0.11.0
beautifulsoup4>=4.13.0
jupyterlab>=4.0.0
notebook>=7.0.0
celery>=5.5.0
redis>=4.5.0
celery[redis]>=5.3.0
nltk==3.9.1
flower>=2.0.0
